fix: properly emulate windows prompt for signal tests. (#609)
